backend/requirements.txt
asgiref==3.3.4
blur-detector==0.0.4
cycler==0.10.0
decorator==4.4.2
Django==3.2.4
djangorestframework==3.12.4
imageio==2.9.0
kiwisolver==1.3.1
matplotlib==3.4.2
networkx==2.5.1
numpy==1.20.3
opencv-python==4.5.2.54
Pillow==8.2.0
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.1
PyWavelets==1.1.1
scikit-image==0.18.1
scipy==1.6.3
six==1.16.0
sqlparse==0.4.1
tifffile==2021.6.6
